requirements.txt
Django==2.0.2
django-filter==1.1.0
django-money==0.12.3
djangorestframework==3.7.7
Markdown==2.6.11
Pillow==5.0.0
py-moneyed==0.7.0
pytz
django-nested-admin==3.0.21
selenium
coreapi
